requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
aiogram>=3.4.1
pymongo>=4.7.0
python-dotenv>=1.0.1
transformers>=4.40.0
torch>=2.2.0
pytz>=2024.1
Jinja2>=3.1.3
python-multipart>=0.0.9
certifi>=2024.2.2
